requirements.txt
streamlit
langchain
faiss-cpu
huggingface_hub
transformers
pypdf